+++ .gitignore
@@ -109,3 +109,6 @@ slo_games/
 exports/
 leagues_data/
 matches_raw_data/
+*.csv
+*.xlsx
+slds/msgbox.vbs
